+++ HISTORY.rst
@@ -10,6 +10,7 @@ Unreleased
 * Refactored temporary file naming
 * Improve error message when handling empty CSV files
 * Fixed "Argument list too long" error (Issue #15)
+* Added --json parameter
 
 0.2.3 (2017-12-09)
 ------------------